feat: add endpoint to retrieve bus position
The code changes add a new endpoint to retrieve the position of a bus on a specific route. The endpoint makes a request to an external API using the provided client ID and client secret. It filters the response to only include buses with a speed greater than 0 and returns the filtered data.